Refactor spherical simulation files for improved readability and clarity

+++ src/spherical/spherical_calc_rel.c
@@ -1,14 +1,14 @@
 #include <math.h>
 #include <stdlib.h>
 
-#include <utils/types.h>
-#include <utils/macros.h>
 #include <spherical/spherical_calc_rel.h>
+#include <utils/macros.h>
+#include <utils/types.h>
 
-long double rel_sphere_calc_r_dot_dot(long double k_sqr, long double Hbar_sqr, long double mass,
-                                 long double gamma, long double r, long double charge,
-                                 long double r_dot)
-{
+long double rel_sphere_calc_r_dot_dot(long double k_sqr, long double Hbar_sqr,
+                                      long double mass, long double gamma,
+                                      long double r, long double charge,
+                                      long double r_dot) {
     long double arg1;
     long double arg2;
     long double arg3;
@@ -34,9 +34,9 @@ long double rel_sphere_calc_r_dot_dot(long double k_sqr, long double Hbar_sqr, l
     return arg1;
 }
 
-long double rel_sphere_calc_phi_dot(long double N_phi, long double Hbar, long double theta,
-                               long double mass, long double r, long double gamma)
-{
+long double rel_sphere_calc_phi_dot(long double N_phi, long double Hbar,
+                                    long double theta, long double mass,
+                                    long double r, long double gamma) {
 
     long double arg1 = N_phi * Hbar;
 
@@ -47,8 +47,8 @@ long double rel_sphere_calc_phi_dot(long double N_phi, long double Hbar, long do
     return arg1 / arg2;
 }
 
-long double rel_sphere_calc_theta_dot_dot(struct spherical_calc_rel_params params)
-{
+long double
+rel_sphere_calc_theta_dot_dot(struct spherical_calc_rel_params params) {
     long double r = params.r;
     long double r_dot = params.r_dot;
     long double theta = params.theta;
@@ -72,10 +72,9 @@ long double rel_sphere_calc_theta_dot_dot(struct spherical_calc_rel_params param
     return arg1 - arg2;
 }
 
-long double rel_sphere_calc_gamma(
-    long double k, long double Hbar, long double mass,
-    long double r, long double r_dot)
-{
+long double rel_sphere_calc_gamma(long double k, long double Hbar,
+                                  long double mass, long double r,
+                                  long double r_dot) {
     long double arg1 = C;
     arg1 *= mass * r;
     arg1 = (k * Hbar) / arg1;
@@ -92,15 +91,14 @@ long double rel_sphere_calc_gamma(
     return arg1;
 }
 
-long double rel_sphere_calc_theta_min(long double N_phi, long double k)
-{
+long double rel_sphere_calc_theta_min(long double N_phi, long double k) {
     return asinl(N_phi / k);
 }
 
-long double rel_sphere_calc_spc_case_r_dot_dot(
-    long double r, long double theta_dot, long double charge,
-    long double mass)
-{
+long double rel_sphere_calc_spc_case_r_dot_dot(long double r,
+                                               long double theta_dot,
+                                               long double charge,
+                                               long double mass) {
 
     long double arg1 = r * theta_dot * theta_dot;
 
@@ -111,10 +109,9 @@ long double rel_sphere_calc_spc_case_r_dot_dot(
     return arg1 - arg2;
 }
 
-long double rel_sphere_calc_spc_case_theta_dot(
-    long double k, long double Hbar, long double mass,
-    long double r, long double gamma)
-{
+long double rel_sphere_calc_spc_case_theta_dot(long double k, long double Hbar,
+                                               long double mass, long double r,
+                                               long double gamma) {
 
     long double arg1 = k * Hbar;
 
@@ -123,8 +120,8 @@ long double rel_sphere_calc_spc_case_theta_dot(
     return arg1 / arg2;
 }
 
-long double rel_sphere_calc_delta_phi(const long double *curr_max_vec, const long double *prev_max_vec)
-{
+long double rel_sphere_calc_delta_phi(const long double *curr_max_vec,
+                                      const long double *prev_max_vec) {
 
     long double arg1 = dot_product(curr_max_vec, prev_max_vec, 3);
 
@@ -135,8 +132,7 @@ long double rel_sphere_calc_delta_phi(const long double *curr_max_vec, const lon
     return acosl(arg1);
 }
 
-long double *stoc(long double roh, long double phi, long double theta)
-{
+long double *stoc(long double roh, long double phi, long double theta) {
 
     long double *cords = malloc(sizeof(*cords) * 3);
 
@@ -151,11 +147,10 @@ long double *stoc(long double roh, long double phi, long double theta)
     return cords;
 }
 
-long double dot_product(const long double *v1, const long double *v2, size_t size)
-{
+long double dot_product(const long double *v1, const long double *v2,
+                        size_t size) {
     long double sum = 0;
-    for (int i = 0; i < size; i++)
-    {
+    for (size_t i = 0; i < size; i++) {
         sum += v1[i] * v2[i];
     }
     return sum;
